+++ SwiftyRSA/SwiftyRSA.swift
@@ -29,14 +29,23 @@ public class SwiftyRSAError: NSError {
    */
 
 @objc
-public class VerificationResult: NSObject, BooleanType {
+public class VerificationResult: NSObject {
+    public let isSuccessful: Bool
+    public let error: NSError?
     
-     /// `true` if the signature was verified
+    init(isSuccessful: Bool, error: NSError?) {
+        self.isSuccessful = isSuccessful
+        self.error = error
+    }
+}
+
+extension CFString: Hashable {
+    public var hashValue: Int {
+        return (self as String).hashValue
+    }
     
-    public let boolValue: Bool
-    
-    init(_ boolValue: Bool) {
-        self.boolValue = boolValue
+    static public func == (lhs: CFString, rhs: CFString) -> Bool {
+        return lhs as String == rhs as String
     }
 }
 
@@ -51,50 +60,50 @@ public class SwiftyRSA: NSObject {
         case SHA512
     }
     
-    private var keyTags: [NSData] = []
+    private var keyTags: [Data] = []
     private static let defaultPadding: SecPadding = .PKCS1
     private static var defaultDigest: DigestType = .SHA1
     
     // MARK: - Public Shorthands
     
-    public class func encryptString(str: String, publicKeyPEM: String, padding: SecPadding = defaultPadding) throws -> String {
+    public class func encryptString(_ str: String, publicKeyPEM: String, padding: SecPadding = defaultPadding) throws -> String {
         let rsa = SwiftyRSA()
         let key = try rsa.publicKeyFromPEMString(publicKeyPEM)
         return try rsa.encryptString(str, publicKey: key, padding: padding)
     }
     
-    public class func encryptString(str: String, publicKeyDER: NSData, padding: SecPadding = defaultPadding) throws -> String {
+    public class func encryptString(_ str: String, publicKeyDER: Data, padding: SecPadding = defaultPadding) throws -> String {
         let rsa = SwiftyRSA()
         let key = try rsa.publicKeyFromDERData(publicKeyDER)
         return try rsa.encryptString(str, publicKey: key, padding: padding)
     }
     
-    public class func decryptString(str: String, privateKeyPEM: String, padding: SecPadding = defaultPadding) throws -> String {
+    public class func decryptString(_ str: String, privateKeyPEM: String, padding: SecPadding = defaultPadding) throws -> String {
         let rsa = SwiftyRSA()
         let key = try rsa.privateKeyFromPEMString(privateKeyPEM)
         return try rsa.decryptString(str, privateKey: key, padding: padding)
     }
     
-    public class func encryptData(data: NSData, publicKeyPEM: String, padding: SecPadding = defaultPadding) throws -> NSData {
+    public class func encryptData(_ data: Data, publicKeyPEM: String, padding: SecPadding = defaultPadding) throws -> Data {
         let rsa = SwiftyRSA()
         let key = try rsa.publicKeyFromPEMString(publicKeyPEM)
         return try rsa.encryptData(data, publicKey: key, padding: padding)
     }
     
-    public class func encryptData(data: NSData, publicKeyDER: NSData, padding: SecPadding = defaultPadding) throws -> NSData {
+    public class func encryptData(_ data: Data, publicKeyDER: Data, padding: SecPadding = defaultPadding) throws -> Data {
         let rsa = SwiftyRSA()
         let key = try rsa.publicKeyFromDERData(publicKeyDER)
         return try rsa.encryptData(data, publicKey: key, padding: padding)
     }
     
-    public class func decryptData(data: NSData, privateKeyPEM: String, padding: SecPadding = defaultPadding) throws -> NSData {
+    public class func decryptData(_ data: Data, privateKeyPEM: String, padding: SecPadding = defaultPadding) throws -> Data {
         let rsa = SwiftyRSA()
         let key = try rsa.privateKeyFromPEMString(privateKeyPEM)
         return try rsa.decryptData(data, privateKey: key, padding: padding)
     }
     
     /**
-     Sign a `String` using a private key.  The supplied string will be hashed using the specified 
+     Sign a `String` using a private key.  The supplied string will be hashed using the specified
      hashing function and the resulting digest will be signed.
      
      - parameter str: The `String` to be signed.
@@ -104,7 +113,7 @@ public class SwiftyRSA: NSObject {
      - throws: `SwiftyRSAError` if there is an error in the signing process
      */
     
-    public class func signString(str: String, privateKeyPEM: String, digestMethod: DigestType = defaultDigest) throws -> String {
+    public class func signString(_ str: String, privateKeyPEM: String, digestMethod: DigestType = defaultDigest) throws -> String {
         let rsa = SwiftyRSA()
         let key = try rsa.privateKeyFromPEMString(privateKeyPEM)
         return try rsa.signString(str, privateKey: key, digestMethod: digestMethod)
@@ -121,7 +130,7 @@ public class SwiftyRSA: NSObject {
      - throws: `SwiftyRSAError` if there is an error in the signing process
      */
     
-    public class func signData(data: NSData, privateKeyPEM: String, digestMethod: DigestType = defaultDigest) throws -> NSData {
+    public class func signData(_ data: Data, privateKeyPEM: String, digestMethod: DigestType = defaultDigest) throws -> Data {
         let rsa = SwiftyRSA()
         let key = try rsa.privateKeyFromPEMString(privateKeyPEM)
         return try rsa.signData(data, privateKey: key, digestMethod: digestMethod)
@@ -139,10 +148,15 @@ public class SwiftyRSA: NSObject {
      - throws: `SwiftyRSAError` if there is an error in the verification process
      */
     
-    public class func verifySignatureString(str: String, signature: String, publicKeyPEM: String, digestMethod: DigestType = defaultDigest) throws -> VerificationResult {
+    public class func verifySignatureString(_ str: String, signature: String, publicKeyPEM: String, digestMethod: DigestType = defaultDigest) -> VerificationResult {
         let rsa = SwiftyRSA()
-        let key = try rsa.publicKeyFromPEMString(publicKeyPEM)
-        return try rsa.verifySignatureString(str, signature: signature, publicKey: key, digestMethod: digestMethod)
+        let key: SecKey
+        do {
+            key = try rsa.publicKeyFromPEMString(publicKeyPEM)
+        } catch {
+            return VerificationResult(isSuccessful: false, error: error as NSError)
+        }
+        return rsa.verifySignatureString(str, signature: signature, publicKey: key, digestMethod: digestMethod)
     }
     
     /**
@@ -157,10 +171,15 @@ public class SwiftyRSA: NSObject {
      - throws: `SwiftyRSAError` if there is an error in the verification process
      */
     
-    public class func verifySignatureData(data: NSData, signature: NSData, publicKeyPEM: String, digestMethod: DigestType = defaultDigest) throws -> VerificationResult {
+    public class func verifySignatureData(_ data: Data, signature: Data, publicKeyPEM: String, digestMethod: DigestType = defaultDigest) -> VerificationResult {
         let rsa = SwiftyRSA()
-        let key = try rsa.publicKeyFromPEMString(publicKeyPEM)
-        return try rsa.verifySignatureData(data, signatureData: signature, publicKey: key, digestMethod: digestMethod)
+        let key: SecKey
+        do {
+            key = try rsa.publicKeyFromPEMString(publicKeyPEM)
+        } catch {
+            return VerificationResult(isSuccessful: false, error: error as NSError)
+        }
+        return rsa.verifySignatureData(data, signatureData: signature, publicKey: key, digestMethod: digestMethod)
     }
     
     /**
@@ -175,10 +194,15 @@ public class SwiftyRSA: NSObject {
      - throws: `SwiftyRSAError` if there is an error in the verification process
      */
     
-    public class func verifySignatureString(str: String, signature: String, publicKeyDER: NSData, digestMethod: DigestType = defaultDigest) throws -> VerificationResult {
+    public class func verifySignatureString(_ str: String, signature: String, publicKeyDER: Data, digestMethod: DigestType = defaultDigest) -> VerificationResult {
         let rsa = SwiftyRSA()
-        let key = try rsa.publicKeyFromDERData(publicKeyDER)
-        return try rsa.verifySignatureString(str, signature: signature, publicKey: key, digestMethod: digestMethod)
+        let key: SecKey
+        do {
+            key = try rsa.publicKeyFromDERData(publicKeyDER)
+        } catch {
+            return VerificationResult(isSuccessful: false, error: error as NSError)
+        }
+        return rsa.verifySignatureString(str, signature: signature, publicKey: key, digestMethod: digestMethod)
     }
     
     /**
@@ -193,10 +217,15 @@ public class SwiftyRSA: NSObject {
      - throws: `SwiftyRSAError` if there is an error in the verification process
      */
     
-    public class func verifySignatureData(data: NSData, signature: NSData, publicKeyDER: NSData, digestMethod: DigestType = defaultDigest) throws -> VerificationResult {
+    public class func verifySignatureData(_ data: Data, signature: Data, publicKeyDER: Data, digestMethod: DigestType = defaultDigest) -> VerificationResult {
         let rsa = SwiftyRSA()
-        let key = try rsa.publicKeyFromDERData(publicKeyDER)
-        return try rsa.verifySignatureData(data, signatureData: signature, publicKey: key, digestMethod: digestMethod)
+        let key: SecKey
+        do {
+            key = try rsa.publicKeyFromDERData(publicKeyDER)
+        } catch {
+            return VerificationResult(isSuccessful: false, error: error as NSError)
+        }
+        return rsa.verifySignatureData(data, signatureData: signature, publicKey: key, digestMethod: digestMethod)
     }
     
 
@@ -206,36 +235,92 @@ public class SwiftyRSA: NSObject {
     	super.init()
     }
     
-    public func publicKeyFromDERData(keyData: NSData) throws -> SecKeyRef {
+    public func publicKeyFromDERData(_ keyData: Data) throws -> SecKey {
         return try addKey(keyData, isPublic: true)
     }
     
-    public func publicKeyFromPEMString(key: String) throws -> SecKeyRef {
+    public func publicKeyFromPEMString(_ key: String) throws -> SecKey {
         let data = try dataFromPEMKey(key)
         return try addKey(data, isPublic: true)
     }
     
-    public func privateKeyFromPEMString(key: String) throws -> SecKeyRef {
+    public func privateKeyFromPEMString(_ key: String) throws -> SecKey {
         let data = try dataFromPEMKey(key)
         return try addKey(data, isPublic: false)
     }
     
+    /** The regular expression used to find public key armor */
+    let publicKeyRegexp : NSRegularExpression? = {
+        let publicKeyRegexp = "(-----BEGIN PUBLIC KEY-----.+?-----END PUBLIC KEY-----)"
+        
+        return try? NSRegularExpression(pattern: publicKeyRegexp, options: .dotMatchesLineSeparators)
+    }()
+    
+    /**
+    Takes an input string, scans for public key sections, and then returns `SecKeyRef`s for any valid keys found
+    
+    - This method scans the file for public key armor - if no keys are found, an empty array is returned
+    - Each public key block found is "parsed" by `publicKeyFromPEMString()` - should that method throw, the error is _swallowed_ and not rethrown
+    
+    This becomes helpful when reading multiple keys in from a single file, or when you have
+    
+    - parameter inputPEMString: The string to use to parse out values
+    - returns: An array of `SecKeyRef` objects
+     
+    - note: This method is marked as `@nonobjc` because NSArray doesn't support storing `SecKeyRef` using generics. If it can be easily exposed to ObjC as is, this can be changed - but currently, cannot be done without wrapping `SecKeyRef`'s which seems circuitous (as this is a fairly Swift'y library).
+    */
+    @nonobjc public func publicKeysFromString(_ inputPEMString:String) -> [SecKey] {
+        var response = [SecKey]()
+        
+        // If our regexp isn't valid, or the input string is empty, we can't move forward…
+        guard let publicKeyRegexp = publicKeyRegexp, inputPEMString.characters.count > 0 else {
+            return response
+        }
+        
+        let all = NSRange(
+            location: 0,
+            length: inputPEMString.characters.count
+        )
+        
+        let matches = publicKeyRegexp.matches(
+            in: inputPEMString,
+            options: NSRegularExpression.MatchingOptions(rawValue: 0),
+            range: all
+        )
+        
+        for result in matches {
+            let match = result.rangeAt(1)
+            let start = inputPEMString.characters.index(inputPEMString.startIndex, offsetBy: match.location)
+            let end = inputPEMString.characters.index(start, offsetBy: match.length)
+            
+            let range = Range<String.Index>(start..<end)
+            
+            let thisKey = inputPEMString[range]
+            
+            if let key = try? self.publicKeyFromPEMString(thisKey) {
+                response.append(key)
+            }
+        }
+        
+        return response
+    }
+    
     // Encrypts data with a RSA key
-    public func encryptData(data: NSData, publicKey: SecKeyRef, padding: SecPadding) throws -> NSData {
+    public func encryptData(_ data: Data, publicKey: SecKey, padding: SecPadding) throws -> Data {
         let blockSize = SecKeyGetBlockSize(publicKey)
-        let maxChunkSize = blockSize - 11
+        let maxChunkSize = (padding == []) ? blockSize : blockSize - 11
         
-        var decryptedDataAsArray = [UInt8](count: data.length / sizeof(UInt8), repeatedValue: 0)
-        data.getBytes(&decryptedDataAsArray, length: data.length)
+        var decryptedDataAsArray = [UInt8](repeating: 0, count: data.count / MemoryLayout<UInt8>.size)
+        (data as NSData).getBytes(&decryptedDataAsArray, length: data.count)
         
-        var encryptedData = [UInt8](count: 0, repeatedValue: 0)
+        var encryptedData = [UInt8](repeating: 0, count: 0)
         var idx = 0
         while (idx < decryptedDataAsArray.count) {
             
             let idxEnd = min(idx + maxChunkSize, decryptedDataAsArray.count)
             let chunkData = [UInt8](decryptedDataAsArray[idx..<idxEnd])
             
-            var encryptedDataBuffer = [UInt8](count: blockSize, repeatedValue: 0)
+            var encryptedDataBuffer = [UInt8](repeating: 0, count: blockSize)
             var encryptedDataLength = blockSize
             
             let status = SecKeyEncrypt(publicKey, padding, chunkData, chunkData.count, &encryptedDataBuffer, &encryptedDataLength)
@@ -249,24 +334,24 @@ public class SwiftyRSA: NSObject {
             idx += maxChunkSize
         }
         
-        return NSData(bytes: encryptedData, length: encryptedData.count)
+        return Data(bytes: UnsafePointer<UInt8>(encryptedData), count: encryptedData.count)
     }
     
     // Decrypt an encrypted data with a RSA key
-    public func decryptData(encryptedData: NSData, privateKey: SecKeyRef, padding: SecPadding) throws -> NSData {
+    public func decryptData(_ encryptedData: Data, privateKey: SecKey, padding: SecPadding) throws -> Data {
         let blockSize = SecKeyGetBlockSize(privateKey)
         
-        var encryptedDataAsArray = [UInt8](count: encryptedData.length / sizeof(UInt8), repeatedValue: 0)
-        encryptedData.getBytes(&encryptedDataAsArray, length: encryptedData.length)
+        var encryptedDataAsArray = [UInt8](repeating: 0, count: encryptedData.count / MemoryLayout<UInt8>.size)
+        (encryptedData as NSData).getBytes(&encryptedDataAsArray, length: encryptedData.count)
         
-        var decryptedData = [UInt8](count: 0, repeatedValue: 0)
+        var decryptedData = [UInt8](repeating: 0, count: 0)
         var idx = 0
         while (idx < encryptedDataAsArray.count) {
             
             let idxEnd = min(idx + blockSize, encryptedDataAsArray.count)
             let chunkData = [UInt8](encryptedDataAsArray[idx..<idxEnd])
             
-            var decryptedDataBuffer = [UInt8](count: blockSize, repeatedValue: 0)
+            var decryptedDataBuffer = [UInt8](repeating: 0, count: blockSize)
             var decryptedDataLength = blockSize
             
             let status = SecKeyDecrypt(privateKey, padding, chunkData, idxEnd-idx, &decryptedDataBuffer, &decryptedDataLength)
@@ -279,25 +364,25 @@ public class SwiftyRSA: NSObject {
             idx += blockSize
         }
         
-        return NSData(bytes: decryptedData, length: decryptedData.count)
+        return Data(bytes: UnsafePointer<UInt8>(decryptedData), count: decryptedData.count)
     }
     
-    public func encryptString(str: String, publicKey: SecKeyRef, padding: SecPadding = defaultPadding) throws -> String {
-        guard let data = str.dataUsingEncoding(NSUTF8StringEncoding) else {
+    public func encryptString(_ str: String, publicKey: SecKey, padding: SecPadding = defaultPadding) throws -> String {
+        guard let data = str.data(using: String.Encoding.utf8) else {
             throw SwiftyRSAError(message: "Couldn't get UT8 data from provided string")
         }
         let encryptedData = try encryptData(data, publicKey: publicKey, padding: padding)
-        return encryptedData.base64EncodedStringWithOptions([])
+        return encryptedData.base64EncodedString(options: [])
     }
     
-    public func decryptString(str: String, privateKey: SecKeyRef, padding: SecPadding = defaultPadding) throws -> String {
-        guard let data =  NSData(base64EncodedString: str, options: []) else {
+    public func decryptString(_ str: String, privateKey: SecKey, padding: SecPadding = defaultPadding) throws -> String {
+        guard let data =  Data(base64Encoded: str, options: []) else {
             throw SwiftyRSAError(message: "Couldn't decode base 64 encoded string")
         }
         
         let decryptedData = try decryptData(data, privateKey: privateKey, padding: padding)
         
-        guard let decryptedString = NSString(data: decryptedData, encoding: NSUTF8StringEncoding) else {
+        guard let decryptedString = NSString(data: decryptedData, encoding: String.Encoding.utf8.rawValue) else {
             throw SwiftyRSAError(message: "Couldn't convert decrypted data to UTF8 string")
         }
         
@@ -319,12 +404,12 @@ public class SwiftyRSA: NSObject {
      - throws: `SwiftyRSAError` if there is an error in the signing process
      */
     
-    public func signString(str: String, privateKey: SecKeyRef, digestMethod: DigestType = defaultDigest) throws -> String {
-        guard let data = str.dataUsingEncoding(NSUTF8StringEncoding) else {
+    public func signString(_ str: String, privateKey: SecKey, digestMethod: DigestType = defaultDigest) throws -> String {
+        guard let data = str.data(using: String.Encoding.utf8) else {
             throw SwiftyRSAError(message: "Couldn't get UTF8 data from provided string")
         }
         let signature = try signData(data, privateKey: privateKey, digestMethod:digestMethod)
-        return signature.base64EncodedStringWithOptions([])
+        return signature.base64EncodedString(options: [])
     }
     
     /**
@@ -338,7 +423,7 @@ public class SwiftyRSA: NSObject {
      - throws: `SwiftyRSAError` if there is an error in the signing process
      */
     
-    public func signData(data: NSData, privateKey: SecKeyRef, digestMethod: DigestType = defaultDigest) throws -> NSData {
+    public func signData(_ data: Data, privateKey: SecKey, digestMethod: DigestType = defaultDigest) throws -> Data {
         
         let (digest, padding) = self.digestForData(data, digestMethod: digestMethod)
         
@@ -354,8 +439,8 @@ public class SwiftyRSA: NSObject {
      - throws: `SwiftyRSAError` if there is an error in the signing process
      */
     
-    @available(*, deprecated = 0.31, message = "Use signDigest() with digestMethod = .SHA1")
-    public func signSHA1Digest(digest: NSData, privateKey: SecKeyRef) throws -> NSData {
+    @available(*, deprecated : 0.31, message : "Use signDigest() with digestMethod = .SHA1")
+    public func signSHA1Digest(_ digest: Data, privateKey: SecKey) throws -> Data {
         
         return try self.signDigest(digest, privateKey: privateKey, padding: .PKCS1SHA1)
         
@@ -371,7 +456,7 @@ public class SwiftyRSA: NSObject {
      - throws: `SwiftyRSAError` if there is an error in the signing process
      */
     
-    public func signDigest(digest: NSData, privateKey: SecKeyRef, digestMethod: DigestType = defaultDigest) throws -> NSData {
+    public func signDigest(_ digest: Data, privateKey: SecKey, digestMethod: DigestType = defaultDigest) throws -> Data {
         
         let (_,padding) = self.digestForData(digest, digestMethod: digestMethod)
         
@@ -393,18 +478,20 @@ public class SwiftyRSA: NSObject {
      - throws: `SwiftyRSAError` if there is an error in the verification process
      */
     
-    public func verifySignatureString(str: String, signature: String, publicKey: SecKeyRef, digestMethod: DigestType = defaultDigest) throws -> VerificationResult {
+    public func verifySignatureString(_ str: String, signature: String, publicKey: SecKey, digestMethod: DigestType = defaultDigest) -> VerificationResult {
         
         
-        guard let data = str.dataUsingEncoding(NSUTF8StringEncoding) else {
-            throw SwiftyRSAError(message: "Couldn't get UTF8 data from provided string")
+        guard let data = str.data(using: String.Encoding.utf8) else {
+            let error = SwiftyRSAError(message: "Couldn't get UTF8 data from provided string")
+            return VerificationResult(isSuccessful: false, error: error)
         }
         
-        guard let signatureData = NSData(base64EncodedString: signature, options: []) else {
-            throw SwiftyRSAError(message: "Couldn't get signature data from provided base64 string")
+        guard let signatureData = Data(base64Encoded: signature, options: []) else {
+            let error = SwiftyRSAError(message: "Couldn't get signature data from provided base64 string")
+            return VerificationResult(isSuccessful: false, error: error)
         }
         
-        return try verifySignatureData(data, signatureData: signatureData, publicKey: publicKey, digestMethod: digestMethod)
+        return verifySignatureData(data, signatureData: signatureData, publicKey: publicKey, digestMethod: digestMethod)
         
     }
     
@@ -417,14 +504,11 @@ public class SwiftyRSA: NSObject {
      - parameter publicKey: A `SecKeyRef` for the public key
      - parameter digestMethod: The `DigestType` that indicates the hashing function
      - returns: A `VerificationResult` that indicates whether the signature was valid or not
-     - throws: `SwiftyRSAError` if there is an error in the verification process
      */
     
-    public func verifySignatureData(data: NSData, signatureData: NSData, publicKey: SecKeyRef, digestMethod: DigestType = defaultDigest) throws -> VerificationResult {
-        
+    public func verifySignatureData(_ data: Data, signatureData: Data, publicKey: SecKey, digestMethod: DigestType = defaultDigest) -> VerificationResult {
         let (digest, padding) = self.digestForData(data, digestMethod: digestMethod)
-        
-        return try verifySignatureData(digest, signature: signatureData, publicKey: publicKey, padding: padding)
+        return verifySignatureData(digest, signature: signatureData, publicKey: publicKey, padding: padding)
         
     }
     
@@ -435,13 +519,11 @@ public class SwiftyRSA: NSObject {
      - parameter signature: The `NSData` containing the signature to be verified.
      - parameter publicKey: A `SecKeyRef` for the public key
      - returns: A `VerificationResult` that indicates whether the signature was valid or not
-     - throws: `SwiftyRSAError` if there is an error in the verification process
      */
     
-    @available(*, deprecated = 0.31, message = "Use verifySignature() with digestMethod = .SHA1")
-    public func verifySHA1SignatureData(SHA1Data: NSData, signature: NSData, publicKey: SecKeyRef) throws -> VerificationResult {
-        
-        return try self.verifySignatureData(SHA1Data, signature: signature, publicKey: publicKey, padding: .PKCS1SHA1)
+    @available(*, deprecated : 0.31, message : "Use verifySignature() with digestMethod = .SHA1")
+    public func verifySHA1SignatureData(_ SHA1Data: Data, signature: Data, publicKey: SecKey) -> VerificationResult {
+        return self.verifySignatureData(SHA1Data, signature: signature, publicKey: publicKey, padding: .PKCS1SHA1)
     }
     
     /**
@@ -455,17 +537,15 @@ public class SwiftyRSA: NSObject {
      - throws: `SwiftyRSAError` if there is an error in the verification process
      */
     
-    public func verifySignatureData(digestData: NSData, signature: NSData, publicKey: SecKeyRef, digestMethod: DigestType = defaultDigest) throws -> VerificationResult {
-        
-        let (_,padding) = self.digestForData(digestData, digestMethod: digestMethod)
-        
-        return try self.verifySignatureData(digestData, signature: signature, publicKey: publicKey, padding: padding)
+    public func verifySignatureData(_ digestData: Data, signature: Data, publicKey: SecKey, digestMethod: DigestType = defaultDigest) -> VerificationResult {
+        let (_, padding) = self.digestForData(digestData, digestMethod: digestMethod)
+        return self.verifySignatureData(digestData, signature: signature, publicKey: publicKey, padding: padding)
     }
     
     
     // MARK: - Private
     
-    private func addKey(keyData: NSData, isPublic: Bool) throws -> SecKeyRef {
+    private func addKey(_ keyData: Data, isPublic: Bool) throws -> SecKey {
         
         var keyData = keyData
         
@@ -474,49 +554,74 @@ public class SwiftyRSA: NSObject {
             try keyData = stripPublicKeyHeader(keyData)
         }
         
-        let tag = NSUUID().UUIDString
-        let tagData = NSData(bytes: tag, length: tag.lengthOfBytesUsingEncoding(NSUTF8StringEncoding))
-        removeKeyWithTagData(tagData)
-        
-        // Add persistent version of the key to system keychain
-        let persistKey = UnsafeMutablePointer<AnyObject?>(nil)
-        let keyClass   = isPublic ? kSecAttrKeyClassPublic : kSecAttrKeyClassPrivate
-        
-        // Add persistent version of the key to system keychain
-        let keyDict = NSMutableDictionary()
-        keyDict.setObject(kSecClassKey,         forKey: kSecClass as! NSCopying)
-        keyDict.setObject(tagData,              forKey: kSecAttrApplicationTag as! NSCopying)
-        keyDict.setObject(kSecAttrKeyTypeRSA,   forKey: kSecAttrKeyType as! NSCopying)
-        keyDict.setObject(keyData,              forKey: kSecValueData as! NSCopying)
-        keyDict.setObject(keyClass,             forKey: kSecAttrKeyClass as! NSCopying)
-        keyDict.setObject(NSNumber(bool: true), forKey: kSecReturnPersistentRef as! NSCopying)
-        keyDict.setObject(kSecAttrAccessibleWhenUnlocked, forKey: kSecAttrAccessible as! NSCopying)
-        
-        var secStatus = SecItemAdd(keyDict as CFDictionary, persistKey)
-        if secStatus != noErr && secStatus != errSecDuplicateItem {
-            throw SwiftyRSAError(message: "Provided key couldn't be added to the keychain")
+        guard let tagData = UUID().uuidString.data(using: .utf8) else {
+            throw SwiftyRSAError(message: "Couldn't create tag data for key")
         }
         
-        keyTags.append(tagData)
+        let keyClass = isPublic ? kSecAttrKeyClassPublic : kSecAttrKeyClassPrivate
         
-        // Now fetch the SecKeyRef version of the key
-        var keyRef: AnyObject? = nil
-        keyDict.removeObjectForKey(kSecValueData)
-        keyDict.removeObjectForKey(kSecReturnPersistentRef)
-        keyDict.setObject(NSNumber(bool: true), forKey: kSecReturnRef as! NSCopying)
-        keyDict.setObject(kSecAttrKeyTypeRSA,   forKey: kSecAttrKeyType as! NSCopying)
-        secStatus = SecItemCopyMatching(keyDict as CFDictionaryRef, &keyRef)
-        
-        guard let unwrappedKeyRef = keyRef else {
-            throw SwiftyRSAError(message: "Couldn't get key reference from the keychain")
+        // On iOS 10+, we can use SecKeyCreateWithData without going through the keychain
+        if #available(iOS 10.0, *) {
+            
+            let sizeInBits = keyData.count * MemoryLayout<UInt8>.size
+            let keyDict: [CFString: Any] = [
+                kSecAttrKeyType: kSecAttrKeyTypeRSA,
+                kSecAttrKeyClass: keyClass,
+                kSecAttrKeySizeInBits: NSNumber(value: sizeInBits)
+            ]
+            
+            guard let key = SecKeyCreateWithData(keyData as CFData, keyDict as CFDictionary, nil) else {
+                throw SwiftyRSAError(message: "Couldn't create key reference from key data")
+            }
+            return key
+            
+        // On iOS 9 and earlier, add a persistent version of the key to the system keychain
+        } else {
+            
+            let persistKey = UnsafeMutablePointer<AnyObject?>(mutating: nil)
+            
+            let keyAddDict: [CFString: Any] = [
+                kSecClass: kSecClassKey,
+                kSecAttrApplicationTag: tagData,
+                kSecAttrKeyType: kSecAttrKeyTypeRSA,
+                kSecValueData: keyData,
+                kSecAttrKeyClass: keyClass,
+                kSecReturnPersistentRef: NSNumber(value: true),
+                kSecAttrAccessible: kSecAttrAccessibleWhenUnlocked
+            ]
+            
+            var secStatus = SecItemAdd(keyAddDict as CFDictionary, persistKey)
+            if secStatus != noErr && secStatus != errSecDuplicateItem {
+                throw SwiftyRSAError(message: "Provided key couldn't be added to the keychain")
+            }
+            
+            // Store the key tag so we can remove it from the keychain later on
+            keyTags.append(tagData)
+            
+            let keyCopyDict: [CFString: Any] = [
+                kSecClass: kSecClassKey,
+                kSecAttrApplicationTag: tagData,
+                kSecAttrKeyType: kSecAttrKeyTypeRSA,
+                kSecAttrKeyClass: keyClass,
+                kSecAttrAccessible: kSecAttrAccessibleWhenUnlocked,
+                kSecReturnRef: NSNumber(value: true),
+            ]
+            
+            // Now fetch the SecKeyRef version of the key
+            var keyRef: AnyObject? = nil
+            secStatus = SecItemCopyMatching(keyCopyDict as CFDictionary, &keyRef)
+            
+            guard let unwrappedKeyRef = keyRef else {
+                throw SwiftyRSAError(message: "Couldn't get key reference from the keychain")
+            }
+            
+            return unwrappedKeyRef as! SecKey
         }
-        
-        return unwrappedKeyRef as! SecKeyRef
     }
     
-    private func dataFromPEMKey(key: String) throws -> NSData {
+    private func dataFromPEMKey(_ key: String) throws -> Data {
         
-        let lines = key.componentsSeparatedByString("\n").filter { line in
+        let lines = key.components(separatedBy: "\n").filter { line in
             return !line.hasPrefix("-----BEGIN") && !line.hasPrefix("-----END")
         }
         
@@ -525,10 +630,10 @@ public class SwiftyRSA: NSObject {
         }
         
         // Decode base64 key
-        let base64EncodedKey = lines.joinWithSeparator("")
-        let keyData = NSData(base64EncodedString: base64EncodedKey, options: .IgnoreUnknownCharacters)
+        let base64EncodedKey = lines.joined(separator: "")
+        let keyData = Data(base64Encoded: base64EncodedKey, options: [])
         
-        guard let unwrappedKeyData = keyData where unwrappedKeyData.length != 0 else {
+        guard let unwrappedKeyData = keyData, unwrappedKeyData.count != 0 else {
             throw SwiftyRSAError(message: "Couldn't decode PEM key data (base64)")
         }
         
@@ -562,15 +667,15 @@ public class SwiftyRSA: NSObject {
      Example of key with X509 header (notice the additional ASN.1 sequence):
      https://lapo.it/asn1js/#30819F300D06092A864886F70D010101050003818D0030818902818100D0674615A252ED3D75D2A3073A0A8A445F3188FD3BEB8BA8584F7299E391BDEC3427F287327414174997D147DD8CA62647427D73C9DA5504E0A3EED5274A1D50A1237D688486FADB8B82061675ABFA5E55B624095DB8790C6DBCAE83D6A8588C9A6635D7CF257ED1EDE18F04217D37908FD0CBB86B2C58D5F762E6207FF7B92D0203010001
      */
-    private func stripPublicKeyHeader(keyData: NSData) throws -> NSData {
-        let count = keyData.length / sizeof(CUnsignedChar)
+    private func stripPublicKeyHeader(_ keyData: Data) throws -> Data {
+        let count = keyData.count / MemoryLayout<CUnsignedChar>.size
         
         guard count > 0 else {
             throw SwiftyRSAError(message: "Provided public key is empty")
         }
         
-        var byteArray = [UInt8](count: count, repeatedValue: 0)
-        keyData.getBytes(&byteArray, length: keyData.length)
+        var byteArray = [UInt8](repeating: 0, count: count)
+        (keyData as NSData).getBytes(&byteArray, length: keyData.count)
         
         var index = 0
         guard byteArray[index] == 0x30 else {
@@ -617,33 +722,33 @@ public class SwiftyRSA: NSObject {
         
         index += 1
         
-        let strippedKeyBytes = [UInt8](byteArray[index...keyData.length - 1])
-        let data = NSData(bytes: strippedKeyBytes, length: keyData.length - index)
+        let strippedKeyBytes = [UInt8](byteArray[index...keyData.count - 1])
+        let data = Data(bytes: UnsafePointer<UInt8>(strippedKeyBytes), count: keyData.count - index)
         
         return data
     }
     
-    private func removeKeyWithTagData(tagData: NSData) {
+    private func removeKeyWithTagData(_ tagData: Data) {
         let publicKey = NSMutableDictionary()
         publicKey.setObject(kSecClassKey,       forKey: kSecClass as! NSCopying)
         publicKey.setObject(kSecAttrKeyTypeRSA, forKey: kSecAttrKeyType as! NSCopying)
         publicKey.setObject(tagData,            forKey: kSecAttrApplicationTag as! NSCopying)
-        SecItemDelete(publicKey as CFDictionaryRef)
+        SecItemDelete(publicKey as CFDictionary)
     }
     
-    private func signDigest(digest: NSData, privateKey: SecKeyRef, padding: SecPadding) throws -> NSData {
+    private func signDigest(_ digest: Data, privateKey: SecKey, padding: SecPadding) throws -> Data {
         
         let blockSize = SecKeyGetBlockSize(privateKey)
         let maxChunkSize = blockSize - 11
         
-        guard (digest.length / sizeof(UInt8) <= maxChunkSize) else {
+        guard (digest.count / MemoryLayout<UInt8>.size <= maxChunkSize) else {
             throw SwiftyRSAError(message: "data length exceeds \(maxChunkSize)")
         }
         
-        var signDataAsArray = [UInt8](count: digest.length / sizeof(UInt8), repeatedValue: 0)
-        digest.getBytes(&signDataAsArray, length: digest.length)
+        var signDataAsArray = [UInt8](repeating: 0, count: digest.count / MemoryLayout<UInt8>.size)
+        (digest as NSData).getBytes(&signDataAsArray, length: digest.count)
         
-        var signatureData = [UInt8](count: blockSize, repeatedValue: 0)
+        var signatureData = [UInt8](repeating: 0, count: blockSize)
         var signatureDataLength = blockSize
         
         let status = SecKeyRawSign(privateKey, padding, signDataAsArray, signDataAsArray.count, &signatureData, &signatureDataLength)
@@ -654,48 +759,49 @@ public class SwiftyRSA: NSObject {
         }
         
         
-        return NSData(bytes: signatureData, length: signatureData.count)
+        return Data(bytes: UnsafePointer<UInt8>(signatureData), count: signatureData.count)
     }
     
-    private func verifySignatureData(SHAData: NSData, signature: NSData, publicKey: SecKeyRef, padding: SecPadding) throws -> VerificationResult {
+    private func verifySignatureData(_ SHAData: Data, signature: Data, publicKey: SecKey, padding: SecPadding) -> VerificationResult {
         
-        var verifyDataAsArray = [UInt8](count: SHAData.length / sizeof(UInt8), repeatedValue: 0)
-        SHAData.getBytes(&verifyDataAsArray, length: SHAData.length)
+        var verifyDataAsArray = [UInt8](repeating: 0, count: SHAData.count / MemoryLayout<UInt8>.size)
+        (SHAData as NSData).getBytes(&verifyDataAsArray, length: SHAData.count)
         
-        var signatureDataAsArray = [UInt8](count: signature.length / sizeof(UInt8), repeatedValue: 0)
-        signature.getBytes(&signatureDataAsArray, length: signature.length)
+        var signatureDataAsArray = [UInt8](repeating: 0, count: signature.count / MemoryLayout<UInt8>.size)
+        (signature as NSData).getBytes(&signatureDataAsArray, length: signature.count)
         
         let status = SecKeyRawVerify(publicKey, padding, verifyDataAsArray, verifyDataAsArray.count, signatureDataAsArray, signatureDataAsArray.count)
         
         if (status == errSecSuccess) {
-            return VerificationResult(true)
+            return VerificationResult(isSuccessful: true, error: nil)
         } else if (status == -9809) {
-            return VerificationResult(false)
+            return VerificationResult(isSuccessful: false, error: nil)
         } else {
-            throw SwiftyRSAError(message: "Couldn't verify signature - \(status)")
+            let error = SwiftyRSAError(message: "Couldn't verify signature - \(status)")
+            return VerificationResult(isSuccessful: false, error: error)
         }
     }
     
-    private func digestForData(data: NSData, digestMethod: DigestType) -> (digest:NSData, padding:SecPadding) {
+    private func digestForData(_ data: Data, digestMethod: DigestType) -> (digest:Data, padding:SecPadding) {
         
-        var digest: NSData
+        var digest: Data
         var padding: SecPadding
         
         switch digestMethod {
         case .SHA1:
-            digest = data.SwiftyRSASHA1()
+            digest = data.swiftyRSASHA1
             padding = .PKCS1SHA1
         case .SHA224:
-            digest = data.SwiftyRSASHA224()
+            digest = data.swiftyRSASHA224
             padding = .PKCS1SHA224
         case .SHA256:
-            digest = data.SwiftyRSASHA256()
+            digest = data.swiftyRSASHA256
             padding = .PKCS1SHA256
         case .SHA384:
-            digest = data.SwiftyRSASHA384()
+            digest = data.swiftyRSASHA384
             padding = .PKCS1SHA384
         case .SHA512:
-            digest = data.SwiftyRSASHA512()
+            digest = data.swiftyRSASHA512
             padding = .PKCS1SHA512
         }