+++ CHANGELOG.md
@@ -2,10 +2,19 @@ SwiftyRSA Changelog
 ===================
 
 # master
+
+ - Don't reduce maxmim blocksize when padding is `None` [#29](https://github.com/TakeScoop/SwiftyRSA/issues/29)
+
+# [0.4.0]
+
  - Add support for SHA2 (224,256,384 & 512 bits) digest algorithms
- - `verifySHA1SignatureData` & `signSHA1Digest` are now deprecated; use `verifySignature()` 
+ - `verifySHA1SignatureData` & `signSHA1Digest` are now deprecated; use `verifySignature()`
    and `signDigest()`
  - Objective-C sign & verification functions now require a `digestMethod:` parameter
+ - Added support to read multiple keys from an input file using `publicKeysFromString()`.
+	[#22](https://github.com/TakeScoop/SwiftyRSA/pull/22)
+ - Added WatchOS and tvOS support.
+   [#23](https://github.com/TakeScoop/SwiftyRSA/pull/23)
 
 # [0.3.0]
 
@@ -37,7 +46,8 @@ SwiftyRSA Changelog
 
 Initial release.
 
-[master]: https://github.com/TakeScoop/SwiftyRSA/compare/0.3.0...master
+[master]: https://github.com/TakeScoop/SwiftyRSA/compare/0.4.0...master
+[0.4.0]: https://github.com/TakeScoop/SwiftyRSA/releases/tag/0.4.0
 [0.3.0]: https://github.com/TakeScoop/SwiftyRSA/releases/tag/0.3.0
 [0.2.1]: https://github.com/TakeScoop/SwiftyRSA/releases/tag/0.2.1
 [0.2.0]: https://github.com/TakeScoop/SwiftyRSA/releases/tag/0.2.0